Move tests, re-export items from core::ascii
